chatbot-training-service2/requirements.txt
fastapi[all]
openai
python-dotenv
pydantic==1.*
langchain
faiss-cpu
tiktoken
flask==3.0.0
flask_cors==4.0.0
sentence-transformers==2.2.2